+++ ConsoleApp/Program.cs
@@ -26,7 +26,7 @@ namespace ConsoleApp
 
             for (int i = 0; i < 7; i++)
             {
-                personlist.Add(RandomPerson.GetRandomAdultOrChild());
+                personlist.Add(RandomPerson.GetRandomAnyPerson());
             }
 
             // Печать исходного списка
@@ -40,20 +40,21 @@ namespace ConsoleApp
 
             switch (person)
             {
-                //TODO: pattern matching
-                case Adult adult:
-                {
-                    Console.WriteLine("This is an adult!");
-                    Console.WriteLine(adult.MythologicalCreature());
-                    break;
-                }
+                case Adult:
+                    {
+                        Adult adult = (Adult)person;
+                        Console.WriteLine("This is an adult!");
+                        Console.WriteLine(adult.MythologicalCreature());
+                        break;
+                    }
 
-                case Child child:
-                {
-                    Console.WriteLine("It's a child!");
-                    Console.WriteLine(child.Dishes());
-                    break;
-                }
+                case Child:
+                    {
+                        Child child = (Child)person;
+                        Console.WriteLine("It's a child!");
+                        Console.WriteLine(child.Dishes());
+                        break;
+                    }
                 default:
                     break;
             }
@@ -62,7 +63,6 @@ namespace ConsoleApp
                 "\n\t\tCreating a new person.");
             _ = Console.ReadKey();
 
-            //TODO: RSDN
             PersonBase newperson = ConsolePerson.InputPersonByConsole();
 
             Console.WriteLine("\nA new person: ");
