+++ Model/Gender.cs
@@ -7,6 +7,7 @@ namespace Model
     public enum Gender
     {
         Male,
-        Female
+        Female,
+        Default
     }
 }
